20241015-223037

+++ app/api/src/users/user.3_service.js
@@ -26,22 +26,36 @@ export class UserService {
         let date = Date.now().toString();
         dbNewUser.passwordHash = createPasswordHash(newUser.password, date);
         dbNewUser.dateCreated = date;
+        dbNewUser.isActive = 1;
+        dbNewUser.isAdmin = 0;
         dbNewUser.token = "";
         let dbCreate = await this.dbService.create(this.table, dbNewUser);
         console.log("dbNewUser", dbCreate);
         if (!dbCreate) {
-            return { create: false, user: {} };
+            return { create: false, error: "User not created" };
         }
         return { create: true, user: DTOUserFromDBToUser(dbNewUser) };
     }
     async update(id, data) {
-        let dbUpdate = this.dbService.update(this.table, id, data);
+        let dbUser = await this.dbService.getOne(this.table, id);
+        if (data.password && data.password !== "") {
+            data.passwordHash = createPasswordHash(
+                data.password,
+                dbUser.dateCreated
+            );
+        }
+        delete data.password;
+        let dbUpdate = await this.dbService.update(this.table, id, data);
         if (!dbUpdate) {
             return { update: false, user: {} };
         }
         return { update: true, user: DTOUserFromDBToUser(data) };
     }
     async delete(id) {
+        const user = await this.getOne(id);
+        if (!user) {
+            return { delete: false, id: id };
+        }
         return this.dbService.delete(this.table, id);
     }
     async createInitUsers() {
@@ -52,6 +66,8 @@ export class UserService {
                 email: "admin@localhost",
                 dateCreated: date,
                 passwordHash: createPasswordHash("admin", date),
+                isActive: 1,
+                isAdmin: 1,
                 token: "",
             };
             console.log(
@@ -86,18 +102,29 @@ const checkPasswords = (password, passwordHash, secret) => {
     return createPasswordHash(password, secret) === passwordHash;
 };
 
-const DTOCreateUserToDBUser = ({ email, passwordHash, dateCreated, token }) => {
+const DTOCreateUserToDBUser = ({
+    email,
+    passwordHash,
+    dateCreated,
+    isActive,
+    isAdmin,
+    token,
+}) => {
     return {
         id: crypto.randomUUID(),
         email,
         passwordHash,
         dateCreated,
+        isActive,
+        isAdmin,
         token,
     };
 };
-const DTOUserFromDBToUser = ({ id, email }) => {
+const DTOUserFromDBToUser = ({ id, email, isActive, isAdmin }) => {
     return {
         id,
         email,
+        isActive,
+        isAdmin,
     };
 };
